requirements.txt
requests
pycrypto